Use util indexer and database from watcher-ts
