+++ packages/erc20-watcher/src/entity/Contract.ts
@@ -16,6 +16,9 @@ export class Contract {
   @Column('varchar', { length: 8 })
   kind!: string;
 
+  @Column('boolean', { default: false })
+  checkpoint!: boolean;
+
   @Column('integer')
   startingBlock!: number;
 }
