+++ packages/uni-info-watcher/src/client.ts
@@ -4,7 +4,7 @@
 
 import { gql } from '@apollo/client/core';
 import { GraphQLClient, GraphQLConfig } from '@vulcanize/ipld-eth-client';
-import { BlockHeight, OrderDirection } from '@vulcanize/util';
+import { BlockHeight, OrderDirection } from '@cerc-io/util';
 
 import {
   queryBundles,
