+++ packages/uni-info-watcher/src/database.ts
@@ -4,25 +4,29 @@
 
 import assert from 'assert';
 import {
+  Brackets,
   Connection,
   ConnectionOptions,
   DeepPartial,
   FindConditions,
   FindManyOptions,
   LessThanOrEqual,
-  QueryRunner
+  QueryRunner,
+  Repository,
+  SelectQueryBuilder
 } from 'typeorm';
+import { RawSqlResultsToEntityTransformer } from 'typeorm/query-builder/transformer/RawSqlResultsToEntityTransformer';
 import path from 'path';
 import { SelectionNode } from 'graphql';
+import debug from 'debug';
+import _ from 'lodash';
 
 import {
-  Database as BaseDatabase,
-  DatabaseInterface,
-  BlockHeight,
-  QueryOptions,
-  Where,
-  ENTITY_QUERY_TYPE
+  eventProcessingLoadEntityCacheHitCount,
+  eventProcessingLoadEntityCount,
+  eventProcessingLoadEntityDBQueryDuration
 } from '@vulcanize/util';
+import { StateKind, Database as BaseDatabase, DatabaseInterface, BlockHeight, QueryOptions, Where } from '@cerc-io/util';
 
 import { Factory } from './entity/Factory';
 import { Pool } from './entity/Pool';
@@ -47,8 +51,30 @@ import { SyncStatus } from './entity/SyncStatus';
 import { TickDayData } from './entity/TickDayData';
 import { Contract } from './entity/Contract';
 import { IPLDBlock } from './entity/IPLDBlock';
-import { StateKind } from '@cerc-io/util';
 import { IpldStatus } from './entity/IpldStatus';
+
+const log = debug('vulcanize:database');
+
+export const DEFAULT_LIMIT = 100;
+
+export enum ENTITY_QUERY_TYPE {
+  SINGULAR,
+  DISTINCT_ON,
+  GROUP_BY
+}
+
+// Cache for updated entities used in job-runner event processing.
+export interface CachedEntities {
+  frothyBlocks: Map<
+    string,
+    {
+      blockNumber: number;
+      parentHash: string;
+      entities: Map<string, Map<string, { [key: string]: any }>>;
+    }
+  >;
+  latestPrunedEntities: Map<string, Map<string, { [key: string]: any }>>;
+}
 
 // Map: Entity to suitable query type.
 const ENTITY_QUERY_TYPE_MAP = new Map<new() => any, number>([
@@ -77,6 +103,11 @@ export class Database implements DatabaseInterface {
   _baseDatabase: BaseDatabase
   _relationsMap: Map<any, { [key: string]: any }>
 
+  _cachedEntities: CachedEntities = {
+    frothyBlocks: new Map(),
+    latestPrunedEntities: new Map()
+  }
+
   constructor (config: ConnectionOptions) {
     assert(config);
 
@@ -95,7 +126,7 @@ export class Database implements DatabaseInterface {
   }
 
   get cachedEntities () {
-    return this._baseDatabase.cachedEntities;
+    return this._cachedEntities;
   }
 
   async init (): Promise<void> {
@@ -179,7 +210,7 @@ export class Database implements DatabaseInterface {
       whereOptions.blockHash = blockHash;
     }
 
-    return this._baseDatabase.getModelEntity(repo, whereOptions);
+    return this.getModelEntity(repo, whereOptions);
   }
 
   async getBundle (queryRunner: QueryRunner, { id, blockHash, blockNumber }: DeepPartial<Bundle>): Promise<Bundle | undefined> {
@@ -194,7 +225,7 @@ export class Database implements DatabaseInterface {
       whereOptions.blockNumber = LessThanOrEqual(blockNumber);
     }
 
-    return this._baseDatabase.getModelEntity(repo, whereOptions);
+    return this.getModelEntity(repo, whereOptions);
   }
 
   async getToken (
@@ -214,14 +245,13 @@ export class Database implements DatabaseInterface {
       whereOptions.blockNumber = LessThanOrEqual(blockNumber);
     }
 
-    let entity = await this._baseDatabase.getModelEntity(repo, whereOptions);
+    let entity = await this.getModelEntity(repo, whereOptions);
 
     if (loadRelations && entity) {
-      [entity] = await this._baseDatabase.loadRelations(
+      [entity] = await this.loadRelations(
         queryRunner,
         { hash: blockHash, number: blockNumber },
         this._relationsMap,
-        ENTITY_QUERY_TYPE_MAP,
         Token,
         [entity],
         selections
@@ -262,14 +292,13 @@ export class Database implements DatabaseInterface {
       whereOptions.blockNumber = LessThanOrEqual(blockNumber);
     }
 
-    let entity = await this._baseDatabase.getModelEntity(repo, whereOptions);
+    let entity = await this.getModelEntity(repo, whereOptions);
 
     if (loadRelations && entity) {
-      [entity] = await this._baseDatabase.loadRelations(
+      [entity] = await this.loadRelations(
         queryRunner,
         { hash: blockHash, number: blockNumber },
         this._relationsMap,
-        ENTITY_QUERY_TYPE_MAP,
         Pool,
         [entity],
         selections
@@ -306,14 +335,13 @@ export class Database implements DatabaseInterface {
         whereOptions.blockHash = blockHash;
       }
 
-      entity = await this._baseDatabase.getModelEntity(repo, whereOptions);
+      entity = await this.getModelEntity(repo, whereOptions);
 
       if (loadRelations && entity) {
-        [entity] = await this._baseDatabase.loadRelations(
+        [entity] = await this.loadRelations(
           queryRunner,
           { hash: blockHash },
           this._relationsMap,
-          ENTITY_QUERY_TYPE_MAP,
           Position,
           [entity]
         );
@@ -333,14 +361,13 @@ export class Database implements DatabaseInterface {
       whereOptions.blockHash = blockHash;
     }
 
-    let entity = await this._baseDatabase.getModelEntity(repo, whereOptions);
+    let entity = await this.getModelEntity(repo, whereOptions);
 
     if (loadRelations && entity) {
-      [entity] = await this._baseDatabase.loadRelations(
+      [entity] = await this.loadRelations(
         queryRunner,
         { hash: blockHash },
         this._relationsMap,
-        ENTITY_QUERY_TYPE_MAP,
         Tick,
         [entity]
       );
@@ -371,14 +398,13 @@ export class Database implements DatabaseInterface {
       whereOptions.blockHash = blockHash;
     }
 
-    let entity = await this._baseDatabase.getModelEntity(repo, whereOptions);
+    let entity = await this.getModelEntity(repo, whereOptions);
 
     if (loadRelations && entity) {
-      [entity] = await this._baseDatabase.loadRelations(
+      [entity] = await this.loadRelations(
         queryRunner,
         { hash: blockHash },
         this._relationsMap,
-        ENTITY_QUERY_TYPE_MAP,
         PoolDayData,
         [entity]
       );
@@ -395,14 +421,13 @@ export class Database implements DatabaseInterface {
       whereOptions.blockHash = blockHash;
     }
 
-    let entity = await this._baseDatabase.getModelEntity(repo, whereOptions);
+    let entity = await this.getModelEntity(repo, whereOptions);
 
     if (loadRelations && entity) {
-      [entity] = await this._baseDatabase.loadRelations(
+      [entity] = await this.loadRelations(
         queryRunner,
         { hash: blockHash },
         this._relationsMap,
-        ENTITY_QUERY_TYPE_MAP,
         PoolHourData,
         [entity]
       );
@@ -419,7 +444,7 @@ export class Database implements DatabaseInterface {
       whereOptions.blockHash = blockHash;
     }
 
-    const entity = await this._baseDatabase.getModelEntity(repo, whereOptions);
+    const entity = await this.getModelEntity(repo, whereOptions);
 
     return entity;
   }
@@ -432,14 +457,13 @@ export class Database implements DatabaseInterface {
       whereOptions.blockHash = blockHash;
     }
 
-    let entity = await this._baseDatabase.getModelEntity(repo, whereOptions);
+    let entity = await this.getModelEntity(repo, whereOptions);
 
     if (loadRelations && entity) {
-      [entity] = await this._baseDatabase.loadRelations(
+      [entity] = await this.loadRelations(
         queryRunner,
         { hash: blockHash },
         this._relationsMap,
-        ENTITY_QUERY_TYPE_MAP,
         TokenDayData,
         [entity]
       );
@@ -456,14 +480,13 @@ export class Database implements DatabaseInterface {
       whereOptions.blockHash = blockHash;
     }
 
-    let entity = await this._baseDatabase.getModelEntity(repo, whereOptions);
+    let entity = await this.getModelEntity(repo, whereOptions);
 
     if (loadRelations && entity) {
-      [entity] = await this._baseDatabase.loadRelations(
+      [entity] = await this.loadRelations(
         queryRunner,
         { hash: blockHash },
         this._relationsMap,
-        ENTITY_QUERY_TYPE_MAP,
         TokenHourData,
         [entity]
       );
@@ -480,14 +503,13 @@ export class Database implements DatabaseInterface {
       whereOptions.blockHash = blockHash;
     }
 
-    let entity = await this._baseDatabase.getModelEntity(repo, whereOptions);
+    let entity = await this.getModelEntity(repo, whereOptions);
 
     if (loadRelations && entity) {
-      [entity] = await this._baseDatabase.loadRelations(
+      [entity] = await this.loadRelations(
         queryRunner,
         { hash: blockHash },
         this._relationsMap,
-        ENTITY_QUERY_TYPE_MAP,
         TickDayData,
         [entity]
       );
@@ -504,7 +526,7 @@ export class Database implements DatabaseInterface {
       whereOptions.blockHash = blockHash;
     }
 
-    const entity = await this._baseDatabase.getModelEntity(repo, whereOptions);
+    const entity = await this.getModelEntity(repo, whereOptions);
 
     return entity;
   }
@@ -521,11 +543,52 @@ export class Database implements DatabaseInterface {
     return entities;
   }
 
-  async getModelEntities<Entity> (queryRunner: QueryRunner, entity: new () => Entity, block: BlockHeight, where: Where = {}, queryOptions: QueryOptions = {}, selections: ReadonlyArray<SelectionNode> = []): Promise<Entity[]> {
-    return this._baseDatabase.getModelEntities(queryRunner, this._relationsMap, ENTITY_QUERY_TYPE_MAP, entity, block, where, queryOptions, selections);
+  async getModelEntities<Entity> (
+    queryRunner: QueryRunner,
+    entity: new () => Entity,
+    block: BlockHeight,
+    where: Where = {},
+    queryOptions: QueryOptions = {},
+    selections: ReadonlyArray<SelectionNode> = []
+  ): Promise<Entity[]> {
+    let entities: Entity[];
+
+    // Use different suitable query patterns based on entities.
+    switch (ENTITY_QUERY_TYPE_MAP.get(entity)) {
+      case ENTITY_QUERY_TYPE.SINGULAR:
+        entities = await this.getModelEntitiesSingular(queryRunner, entity, block, where);
+        break;
+
+      case ENTITY_QUERY_TYPE.DISTINCT_ON:
+        entities = await this.getModelEntitiesDistinctOn(queryRunner, entity, block, where, queryOptions);
+        break;
+
+      case ENTITY_QUERY_TYPE.GROUP_BY:
+        entities = await this.getModelEntitiesGroupBy(queryRunner, entity, block, where, queryOptions);
+        break;
+
+      default:
+        log(`Invalid entity query type for entity ${entity}`);
+        entities = [];
+        break;
+    }
+
+    if (!entities.length) {
+      return [];
+    }
+
+    entities = await this.loadRelations(queryRunner, block, this._relationsMap, entity, entities, selections);
+
+    return entities;
   }
 
-  async getModelEntitiesNoTx<Entity> (entity: new () => Entity, block: BlockHeight, where: Where = {}, queryOptions: QueryOptions = {}, selections: ReadonlyArray<SelectionNode> = []): Promise<Entity[]> {
+  async getModelEntitiesNoTx<Entity> (
+    entity: new () => Entity,
+    block: BlockHeight,
+    where: Where = {},
+    queryOptions: QueryOptions = {},
+    selections: ReadonlyArray<SelectionNode> = []
+  ): Promise<Entity[]> {
     const queryRunner = this._conn.createQueryRunner();
     let res;
 
@@ -539,12 +602,407 @@ export class Database implements DatabaseInterface {
     return res;
   }
 
+  async getModelEntitiesGroupBy<Entity> (
+    queryRunner: QueryRunner,
+    entity: new () => Entity,
+    block: BlockHeight,
+    where: Where = {},
+    queryOptions: QueryOptions = {}
+  ): Promise<Entity[]> {
+    const repo = queryRunner.manager.getRepository(entity);
+    const { tableName } = repo.metadata;
+
+    let subQuery = repo.createQueryBuilder('subTable')
+      .select('subTable.id', 'id')
+      .addSelect('MAX(subTable.block_number)', 'block_number')
+      .addFrom('block_progress', 'blockProgress')
+      .where('subTable.block_hash = blockProgress.block_hash')
+      .andWhere('blockProgress.is_pruned = :isPruned', { isPruned: false })
+      .groupBy('subTable.id');
+
+    if (block.hash) {
+      const { canonicalBlockNumber, blockHashes } = await this._baseDatabase.getFrothyRegion(queryRunner, block.hash);
+
+      subQuery = subQuery
+        .andWhere(new Brackets(qb => {
+          qb.where('subTable.block_hash IN (:...blockHashes)', { blockHashes })
+            .orWhere('subTable.block_number <= :canonicalBlockNumber', { canonicalBlockNumber });
+        }));
+    }
+
+    if (block.number) {
+      subQuery = subQuery.andWhere('subTable.block_number <= :blockNumber', { blockNumber: block.number });
+    }
+
+    let selectQueryBuilder = repo.createQueryBuilder(tableName)
+      .innerJoin(
+        `(${subQuery.getQuery()})`,
+        'latestEntities',
+        `${tableName}.id = "latestEntities"."id" AND ${tableName}.block_number = "latestEntities"."block_number"`
+      )
+      .setParameters(subQuery.getParameters());
+
+    selectQueryBuilder = this._baseDatabase.buildQuery(repo, selectQueryBuilder, where);
+
+    if (queryOptions.orderBy) {
+      selectQueryBuilder = this._baseDatabase.orderQuery(repo, selectQueryBuilder, queryOptions);
+    }
+
+    selectQueryBuilder = this._baseDatabase.orderQuery(repo, selectQueryBuilder, { ...queryOptions, orderBy: 'id' });
+
+    if (queryOptions.skip) {
+      selectQueryBuilder = selectQueryBuilder.offset(queryOptions.skip);
+    }
+
+    if (queryOptions.limit) {
+      selectQueryBuilder = selectQueryBuilder.limit(queryOptions.limit);
+    }
+
+    const entities = await selectQueryBuilder.getMany();
+
+    return entities;
+  }
+
+  async getModelEntitiesDistinctOn<Entity> (
+    queryRunner: QueryRunner,
+    entity: new () => Entity,
+    block: BlockHeight,
+    where: Where = {},
+    queryOptions: QueryOptions = {}
+  ): Promise<Entity[]> {
+    const repo = queryRunner.manager.getRepository(entity);
+
+    let subQuery = repo.createQueryBuilder('subTable')
+      .distinctOn(['subTable.id'])
+      .addFrom('block_progress', 'blockProgress')
+      .where('subTable.block_hash = blockProgress.block_hash')
+      .andWhere('blockProgress.is_pruned = :isPruned', { isPruned: false })
+      .addOrderBy('subTable.id', 'ASC')
+      .addOrderBy('subTable.block_number', 'DESC');
+
+    if (block.hash) {
+      const { canonicalBlockNumber, blockHashes } = await this._baseDatabase.getFrothyRegion(queryRunner, block.hash);
+
+      subQuery = subQuery
+        .andWhere(new Brackets(qb => {
+          qb.where('subTable.block_hash IN (:...blockHashes)', { blockHashes })
+            .orWhere('subTable.block_number <= :canonicalBlockNumber', { canonicalBlockNumber });
+        }));
+    }
+
+    if (block.number) {
+      subQuery = subQuery.andWhere('subTable.block_number <= :blockNumber', { blockNumber: block.number });
+    }
+
+    subQuery = this._baseDatabase.buildQuery(repo, subQuery, where);
+
+    let selectQueryBuilder = queryRunner.manager.createQueryBuilder()
+      .from(
+        `(${subQuery.getQuery()})`,
+        'latestEntities'
+      )
+      .setParameters(subQuery.getParameters());
+
+    if (queryOptions.orderBy) {
+      selectQueryBuilder = this._baseDatabase.orderQuery(repo, selectQueryBuilder, queryOptions, 'subTable_');
+      if (queryOptions.orderBy !== 'id') {
+        selectQueryBuilder = this._baseDatabase.orderQuery(repo, selectQueryBuilder, { ...queryOptions, orderBy: 'id' }, 'subTable_');
+      }
+    }
+
+    if (queryOptions.skip) {
+      selectQueryBuilder = selectQueryBuilder.offset(queryOptions.skip);
+    }
+
+    if (queryOptions.limit) {
+      selectQueryBuilder = selectQueryBuilder.limit(queryOptions.limit);
+    }
+
+    let entities = await selectQueryBuilder.getRawMany();
+    entities = await this._transformResults(queryRunner, repo.createQueryBuilder('subTable'), entities);
+
+    return entities as Entity[];
+  }
+
+  async getModelEntitiesSingular<Entity> (
+    queryRunner: QueryRunner,
+    entity: new () => Entity,
+    block: BlockHeight,
+    where: Where = {}
+  ): Promise<Entity[]> {
+    const repo = queryRunner.manager.getRepository(entity);
+    const { tableName } = repo.metadata;
+
+    let selectQueryBuilder = repo.createQueryBuilder(tableName)
+      .addFrom('block_progress', 'blockProgress')
+      .where(`${tableName}.block_hash = blockProgress.block_hash`)
+      .andWhere('blockProgress.is_pruned = :isPruned', { isPruned: false })
+      .addOrderBy(`${tableName}.block_number`, 'DESC')
+      .limit(1);
+
+    if (block.hash) {
+      const { canonicalBlockNumber, blockHashes } = await this._baseDatabase.getFrothyRegion(queryRunner, block.hash);
+
+      selectQueryBuilder = selectQueryBuilder
+        .andWhere(new Brackets(qb => {
+          qb.where(`${tableName}.block_hash IN (:...blockHashes)`, { blockHashes })
+            .orWhere(`${tableName}.block_number <= :canonicalBlockNumber`, { canonicalBlockNumber });
+        }));
+    }
+
+    if (block.number) {
+      selectQueryBuilder = selectQueryBuilder.andWhere(`${tableName}.block_number <= :blockNumber`, { blockNumber: block.number });
+    }
+
+    selectQueryBuilder = this._baseDatabase.buildQuery(repo, selectQueryBuilder, where);
+
+    const entities = await selectQueryBuilder.getMany();
+
+    return entities as Entity[];
+  }
+
+  async getModelEntity<Entity> (repo: Repository<Entity>, whereOptions: any): Promise<Entity | undefined> {
+    eventProcessingLoadEntityCount.inc();
+
+    const findOptions = {
+      where: whereOptions,
+      order: {
+        blockNumber: 'DESC'
+      }
+    };
+
+    if (findOptions.where.blockHash) {
+      // Check cache only if latestPrunedEntities is updated.
+      // latestPrunedEntities is updated when frothyBlocks is filled till canonical block height.
+      if (this._cachedEntities.latestPrunedEntities.size > 0) {
+        let frothyBlock = this._cachedEntities.frothyBlocks.get(findOptions.where.blockHash);
+        let canonicalBlockNumber = -1;
+
+        // Loop through frothy region until latest entity is found.
+        while (frothyBlock) {
+          const entity = frothyBlock.entities
+            .get(repo.metadata.tableName)
+            ?.get(findOptions.where.id);
+
+          if (entity) {
+            eventProcessingLoadEntityCacheHitCount.inc();
+            return _.cloneDeep(entity) as Entity;
+          }
+
+          canonicalBlockNumber = frothyBlock.blockNumber + 1;
+          frothyBlock = this._cachedEntities.frothyBlocks.get(frothyBlock.parentHash);
+        }
+
+        // Canonical block number is not assigned if blockHash does not exist in frothy region.
+        // Get latest pruned entity from cache only if blockHash exists in frothy region.
+        // i.e. Latest entity in cache is the version before frothy region.
+        if (canonicalBlockNumber > -1) {
+          // If entity not found in frothy region get latest entity in the pruned region.
+          // Check if latest entity is cached in pruned region.
+          const entity = this._cachedEntities.latestPrunedEntities
+            .get(repo.metadata.tableName)
+            ?.get(findOptions.where.id);
+
+          if (entity) {
+            eventProcessingLoadEntityCacheHitCount.inc();
+            return _.cloneDeep(entity) as Entity;
+          }
+
+          // Get latest pruned entity from DB if not found in cache.
+          const endTimer = eventProcessingLoadEntityDBQueryDuration.startTimer();
+          const dbEntity = await this._baseDatabase.getLatestPrunedEntity(repo, findOptions.where.id, canonicalBlockNumber);
+          endTimer();
+
+          if (dbEntity) {
+            // Update latest pruned entity in cache.
+            this.cacheUpdatedEntity(repo, dbEntity, true);
+          }
+
+          return dbEntity;
+        }
+      }
+
+      const endTimer = eventProcessingLoadEntityDBQueryDuration.startTimer();
+      const dbEntity = await this._baseDatabase.getPrevEntityVersion(repo.queryRunner!, repo, findOptions);
+      endTimer();
+
+      return dbEntity;
+    }
+
+    return repo.findOne(findOptions);
+  }
+
+  async loadRelations<Entity> (
+    queryRunner: QueryRunner,
+    block: BlockHeight,
+    relationsMap: Map<any, { [key: string]: any }>,
+    entity: new () => Entity,
+    entities: Entity[],
+    selections: ReadonlyArray<SelectionNode> = []
+  ): Promise<Entity[]> {
+    const relations = relationsMap.get(entity);
+
+    if (!relations) {
+      return entities;
+    }
+
+    // Filter selections from GQL query which are relations.
+    const relationPromises = selections.filter((selection) => selection.kind === 'Field' && Boolean(relations[selection.name.value]))
+      .map(async (selection) => {
+        assert(selection.kind === 'Field');
+        const field = selection.name.value;
+        const { entity: relationEntity, type, foreignKey } = relations[field];
+        let childSelections = selection.selectionSet?.selections || [];
+
+        // Filter out __typename field in GQL for loading relations.
+        childSelections = childSelections.filter(selection => !(selection.kind === 'Field' && selection.name.value === '__typename'));
+
+        switch (type) {
+          case 'one-to-many': {
+            assert(foreignKey);
+
+            const where: Where = {
+              [foreignKey]: [{
+                value: entities.map((entity: any) => entity.id),
+                not: false,
+                operator: 'in'
+              }]
+            };
+
+            const relatedEntities = await this.getModelEntities(
+              queryRunner,
+              relationEntity,
+              block,
+              where,
+              {},
+              childSelections
+            );
+
+            const relatedEntitiesMap = relatedEntities.reduce((acc: {[key:string]: any[]}, entity: any) => {
+              // Related entity might be loaded with data.
+              const parentEntityId = entity[foreignKey].id ?? entity[foreignKey];
+
+              if (!acc[parentEntityId]) {
+                acc[parentEntityId] = [];
+              }
+
+              if (acc[parentEntityId].length < DEFAULT_LIMIT) {
+                acc[parentEntityId].push(entity);
+              }
+
+              return acc;
+            }, {});
+
+            entities.forEach((entity: any) => {
+              if (relatedEntitiesMap[entity.id]) {
+                entity[field] = relatedEntitiesMap[entity.id];
+              } else {
+                entity[field] = [];
+              }
+            });
+
+            break;
+          }
+
+          case 'many-to-many': {
+            const relatedIds = entities.reduce((acc, entity: any) => {
+              entity[field].forEach((relatedEntityId: any) => acc.add(relatedEntityId));
+
+              return acc;
+            }, new Set());
+
+            const where: Where = {
+              id: [{
+                value: Array.from(relatedIds),
+                not: false,
+                operator: 'in'
+              }]
+            };
+
+            const relatedEntities = await this.getModelEntities(
+              queryRunner,
+              relationEntity,
+              block,
+              where,
+              {},
+              childSelections
+            );
+
+            entities.forEach((entity: any) => {
+              const relatedEntityIds: Set<string> = entity[field].reduce((acc: Set<string>, id: string) => {
+                acc.add(id);
+
+                return acc;
+              }, new Set());
+
+              entity[field] = [];
+
+              relatedEntities.forEach((relatedEntity: any) => {
+                if (relatedEntityIds.has(relatedEntity.id) && entity[field].length < DEFAULT_LIMIT) {
+                  entity[field].push(relatedEntity);
+                }
+              });
+            });
+
+            break;
+          }
+
+          default: {
+            // For one-to-one/many-to-one relations.
+            if (childSelections.length === 1 && childSelections[0].kind === 'Field' && childSelections[0].name.value === 'id') {
+              // Avoid loading relation if selections only has id field.
+              entities.forEach((entity: any) => {
+                entity[field] = { id: entity[field] };
+              });
+
+              break;
+            }
+
+            const where: Where = {
+              id: [{
+                value: entities.map((entity: any) => entity[field]),
+                not: false,
+                operator: 'in'
+              }]
+            };
+
+            const relatedEntities = await this.getModelEntities(
+              queryRunner,
+              relationEntity,
+              block,
+              where,
+              {},
+              childSelections
+            );
+
+            const relatedEntitiesMap = relatedEntities.reduce((acc: {[key:string]: any}, entity: any) => {
+              acc[entity.id] = entity;
+
+              return acc;
+            }, {});
+
+            entities.forEach((entity: any) => {
+              if (relatedEntitiesMap[entity[field]]) {
+                entity[field] = relatedEntitiesMap[entity[field]];
+              }
+            });
+
+            break;
+          }
+        }
+      });
+
+    await Promise.all(relationPromises);
+
+    return entities;
+  }
+
   async saveFactory (queryRunner: QueryRunner, factory: Factory, block: Block): Promise<Factory> {
     const repo = queryRunner.manager.getRepository(Factory);
     factory.blockNumber = block.number;
     factory.blockHash = block.hash;
     const data = await repo.save(factory);
-    this._baseDatabase.cacheUpdatedEntity(repo, data);
+    this.cacheUpdatedEntity(repo, data);
 
     return data;
   }
@@ -554,7 +1012,7 @@ export class Database implements DatabaseInterface {
     bundle.blockNumber = block.number;
     bundle.blockHash = block.hash;
     const data = await repo.save(bundle);
-    this._baseDatabase.cacheUpdatedEntity(repo, data);
+    this.cacheUpdatedEntity(repo, data);
 
     return data;
   }
@@ -564,7 +1022,7 @@ export class Database implements DatabaseInterface {
     pool.blockNumber = block.number;
     pool.blockHash = block.hash;
     const data = await repo.save(pool);
-    this._baseDatabase.cacheUpdatedEntity(repo, data);
+    this.cacheUpdatedEntity(repo, data);
 
     return data;
   }
@@ -574,7 +1032,7 @@ export class Database implements DatabaseInterface {
     poolDayData.blockNumber = block.number;
     poolDayData.blockHash = block.hash;
     const data = await repo.save(poolDayData);
-    this._baseDatabase.cacheUpdatedEntity(repo, data);
+    this.cacheUpdatedEntity(repo, data);
 
     return data;
   }
@@ -584,7 +1042,7 @@ export class Database implements DatabaseInterface {
     poolHourData.blockNumber = block.number;
     poolHourData.blockHash = block.hash;
     const data = await repo.save(poolHourData);
-    this._baseDatabase.cacheUpdatedEntity(repo, data);
+    this.cacheUpdatedEntity(repo, data);
 
     return data;
   }
@@ -594,7 +1052,7 @@ export class Database implements DatabaseInterface {
     token.blockNumber = block.number;
     token.blockHash = block.hash;
     const data = await repo.save(token);
-    this._baseDatabase.cacheUpdatedEntity(repo, data);
+    this.cacheUpdatedEntity(repo, data);
 
     return data;
   }
@@ -604,7 +1062,7 @@ export class Database implements DatabaseInterface {
     transaction.blockNumber = block.number;
     transaction.blockHash = block.hash;
     const data = await repo.save(transaction);
-    this._baseDatabase.cacheUpdatedEntity(repo, data);
+    this.cacheUpdatedEntity(repo, data);
 
     return data;
   }
@@ -614,7 +1072,7 @@ export class Database implements DatabaseInterface {
     uniswapDayData.blockNumber = block.number;
     uniswapDayData.blockHash = block.hash;
     const data = await repo.save(uniswapDayData);
-    this._baseDatabase.cacheUpdatedEntity(repo, data);
+    this.cacheUpdatedEntity(repo, data);
 
     return data;
   }
@@ -624,7 +1082,7 @@ export class Database implements DatabaseInterface {
     tokenDayData.blockNumber = block.number;
     tokenDayData.blockHash = block.hash;
     const data = await repo.save(tokenDayData);
-    this._baseDatabase.cacheUpdatedEntity(repo, data);
+    this.cacheUpdatedEntity(repo, data);
 
     return data;
   }
@@ -634,7 +1092,7 @@ export class Database implements DatabaseInterface {
     tokenHourData.blockNumber = block.number;
     tokenHourData.blockHash = block.hash;
     const data = await repo.save(tokenHourData);
-    this._baseDatabase.cacheUpdatedEntity(repo, data);
+    this.cacheUpdatedEntity(repo, data);
 
     return data;
   }
@@ -644,7 +1102,7 @@ export class Database implements DatabaseInterface {
     tick.blockNumber = block.number;
     tick.blockHash = block.hash;
     const data = await repo.save(tick);
-    this._baseDatabase.cacheUpdatedEntity(repo, data);
+    this.cacheUpdatedEntity(repo, data);
 
     return data;
   }
@@ -654,7 +1112,7 @@ export class Database implements DatabaseInterface {
     tickDayData.blockNumber = block.number;
     tickDayData.blockHash = block.hash;
     const data = await repo.save(tickDayData);
-    this._baseDatabase.cacheUpdatedEntity(repo, data);
+    this.cacheUpdatedEntity(repo, data);
 
     return data;
   }
@@ -664,7 +1122,7 @@ export class Database implements DatabaseInterface {
     position.blockNumber = block.number;
     position.blockHash = block.hash;
     const data = await repo.save(position);
-    this._baseDatabase.cacheUpdatedEntity(repo, data);
+    this.cacheUpdatedEntity(repo, data);
 
     return data;
   }
@@ -674,7 +1132,7 @@ export class Database implements DatabaseInterface {
     positionSnapshot.blockNumber = block.number;
     positionSnapshot.blockHash = block.hash;
     const data = await repo.save(positionSnapshot);
-    this._baseDatabase.cacheUpdatedEntity(repo, data);
+    this.cacheUpdatedEntity(repo, data);
 
     return data;
   }
@@ -684,7 +1142,7 @@ export class Database implements DatabaseInterface {
     mint.blockNumber = block.number;
     mint.blockHash = block.hash;
     const data = await repo.save(mint);
-    this._baseDatabase.cacheUpdatedEntity(repo, data);
+    this.cacheUpdatedEntity(repo, data);
 
     return data;
   }
@@ -694,7 +1152,7 @@ export class Database implements DatabaseInterface {
     burn.blockNumber = block.number;
     burn.blockHash = block.hash;
     const data = await repo.save(burn);
-    this._baseDatabase.cacheUpdatedEntity(repo, data);
+    this.cacheUpdatedEntity(repo, data);
 
     return data;
   }
@@ -704,7 +1162,7 @@ export class Database implements DatabaseInterface {
     swap.blockNumber = block.number;
     swap.blockHash = block.hash;
     const data = await repo.save(swap);
-    this._baseDatabase.cacheUpdatedEntity(repo, data);
+    this.cacheUpdatedEntity(repo, data);
 
     return data;
   }
@@ -834,12 +1292,52 @@ export class Database implements DatabaseInterface {
     return this._baseDatabase.removeEntities(queryRunner, entity, findConditions);
   }
 
+  async deleteEntitiesByConditions<Entity> (queryRunner: QueryRunner, entity: new () => Entity, findConditions: FindConditions<Entity>): Promise<void> {
+    await this._baseDatabase.deleteEntitiesByConditions(queryRunner, entity, findConditions);
+  }
+
   async isEntityEmpty<Entity> (entity: new () => Entity): Promise<boolean> {
     return this._baseDatabase.isEntityEmpty(entity);
   }
 
   async getAncestorAtDepth (blockHash: string, depth: number): Promise<string> {
     return this._baseDatabase.getAncestorAtDepth(blockHash, depth);
+  }
+
+  cacheUpdatedEntity<Entity> (repo: Repository<Entity>, entity: any, pruned = false): void {
+    const tableName = repo.metadata.tableName;
+    if (pruned) {
+      let entityIdMap = this._cachedEntities.latestPrunedEntities.get(tableName);
+      if (!entityIdMap) {
+        entityIdMap = new Map();
+      }
+      entityIdMap.set(entity.id, _.cloneDeep(entity));
+      this._cachedEntities.latestPrunedEntities.set(tableName, entityIdMap);
+      return;
+    }
+    const frothyBlock = this._cachedEntities.frothyBlocks.get(entity.blockHash);
+    // Update frothyBlock only if already present in cache.
+    // Might not be present when event processing starts without block processing on job retry.
+    if (frothyBlock) {
+      let entityIdMap = frothyBlock.entities.get(tableName);
+      if (!entityIdMap) {
+        entityIdMap = new Map();
+      }
+      entityIdMap.set(entity.id, _.cloneDeep(entity));
+      frothyBlock.entities.set(tableName, entityIdMap);
+    }
+  }
+
+  async _transformResults<Entity> (queryRunner: QueryRunner, qb: SelectQueryBuilder<Entity>, rawResults: any[]): Promise<any[]> {
+    const transformer = new RawSqlResultsToEntityTransformer(
+      qb.expressionMap,
+      queryRunner.manager.connection.driver,
+      [],
+      [],
+      queryRunner
+    );
+    assert(qb.expressionMap.mainAlias);
+    return transformer.transform(rawResults, qb.expressionMap.mainAlias);
   }
 
   _populateRelationsMap (): void {
