+++ packages/uni-watcher/src/entity/SyncStatus.ts
@@ -34,4 +34,12 @@ export class SyncStatus implements SyncStatusInterface {
 
   @Column('integer')
   latestCanonicalBlockNumber!: number;
+
+  // TODO: Remove default.
+  @Column('varchar', { length: 66, default: '0xf1156cb7e1a2acaa348c66fdc45b02f2547024ded63350a96a46ed12c62b4ff4' })
+  initialIndexedBlockHash!: string;
+
+  // TODO: Remove default.
+  @Column('integer', { default: 12369621 })
+  initialIndexedBlockNumber!: number;
 }
