+++ packages/uni-watcher/src/smoke.test.ts
@@ -22,7 +22,7 @@ import {
   NFPM_ABI
 } from '@vulcanize/util/test';
 import { getCache } from '@vulcanize/cache';
-import { EthClient } from '@vulcanize/ipld-eth-client';
+import { EthClient } from '@cerc-io/ipld-eth-client';
 import {
   abi as FACTORY_ABI
 } from '@uniswap/v3-core/artifacts/contracts/UniswapV3Factory.sol/UniswapV3Factory.json';
@@ -127,7 +127,7 @@ describe('uni-watcher', () => {
     factory = new Contract(factoryContract.address, FACTORY_ABI, signer);
 
     // Verifying with the db.
-    const indexer = new Indexer(db, ethClient, ethProvider, jobQueue);
+    const indexer = new Indexer(config.server, db, ethClient, ethProvider, jobQueue);
     await indexer.init();
     assert(await indexer.isWatchedContract(factory.address), 'Factory contract not added to the database.');
   });
@@ -263,7 +263,7 @@ describe('uni-watcher', () => {
     nfpm = new Contract(nfpmContract.address, NFPM_ABI, signer);
 
     // Verifying with the db.
-    const indexer = new Indexer(db, ethClient, ethProvider, jobQueue);
+    const indexer = new Indexer(config.server, db, ethClient, ethProvider, jobQueue);
     await indexer.init();
     assert(await indexer.isWatchedContract(nfpm.address), 'NFPM contract not added to the database.');
   });
