+++ packages/util/index.ts
@@ -3,14 +3,12 @@
 //
 
 export * from './src/config';
-export * from './src/database';
 export * from './src/job-queue';
 export * from './src/constants';
 export * from './src/misc';
 export * from './src/fill';
 export * from './src/events';
 export * from './src/types';
-export * from './src/indexer';
 export * from './src/job-runner';
 export * from './src/graph-decimal';
 export * from './src/metrics';
